Ruff sweep across src/ tests/ scripts/. 42 errors -> 0. Hand-fixes: SIM108 ternaries (decay.py, measure.py, report.py), SIM103 bare return (security.py), RUF046 redundant int() (hub_exclusion.py), B017 FrozenInstanceError vs Exception (test_dedup_semantic.py, test_graph_edges.py), RUF002/RUF001 ambiguous chars (math_recall.py: ascii doc + per-line noqa on regexes that genuinely match both hyphen and Unicode minus). Auto: import sorts, __slots__ + __all__ ordering, unused noqa, unused imports, unused unpacks (_remapped, _extracted), F541 empty f-strings.
284 tests still pass.

+++ src/nuthatch/clustering/hub_exclusion.py
@@ -59,7 +59,7 @@ def core_nodes(
     }
     sorted_degrees = sorted(degrees.values())
     # Percentile cutoff via linear interpolation rank.
-    rank = int(round((percentile / 100.0) * (len(sorted_degrees) - 1)))
+    rank = round((percentile / 100.0) * (len(sorted_degrees) - 1))
     cutoff = sorted_degrees[rank]
     return [node for node, d in degrees.items() if d >= cutoff]
